Replace redundant tests with robust well check test
