+++ src/common/mod.rs
@@ -1,5 +1,4 @@
 pub mod auth;
-pub mod errors;
 pub mod models;
 pub mod state;
