+++ src/routes/assets/models.rs
@@ -1,7 +1,7 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
 
 use crate::external::s3::delete_from_s3;
-use chrono::{DateTime, FixedOffset, Utc};
+use chrono::{DateTime, Utc};
 use crudcrate::{CRUDResource, EntityToModels};
 use sea_orm::entity::prelude::*;
 
@@ -110,7 +110,7 @@ async fn delete_many_assets(db: &DatabaseConnection, ids: Vec<Uuid>) -> Result<V
     }
 
     // Proceed with deleting the database records
-    let res = Entity::delete_many()
+    let _res = Entity::delete_many()
         .filter(Column::Id.is_in(ids.clone()))
         .exec(db)
         .await?;
