--- src/routes/experiments/comprehensive_validation_tests.rs
@@ -1,1860 +1,0 @@
-use crate::config::test_helpers::setup_test_app;
-use axum::{
-    Router,
-    body::Body,
-    http::{Request, StatusCode},
-};
-use serde_json::{Value, json};
-use std::collections::HashMap;
-use std::fs;
-use tower::ServiceExt;
-
-/// Comprehensive validation data derived from merged.csv analysis
-/// This represents the exact phase transitions that should occur when processing the Excel file
-pub struct WellTransitionData {
-    pub tray: &'static str,        // "P1" or "P2"
-    pub coordinate: &'static str,  // "A1", "B2", etc.
-    pub freeze_time: &'static str, // "2025-03-20 16:19:47"
-    pub temp_probe_1: f64,         // Temperature at freeze time
-    pub row_in_csv: usize,         // Original CSV row for debugging
-}
-
-/// Expected phase transitions extracted from merged.csv analysis
-/// Complete list of all 192 well transitions with exact timestamps and average temperatures
-/// Data generated from systematic CSV analysis of all phase changes (0→1)
-pub const EXPECTED_TRANSITIONS: &[WellTransitionData] = &[
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A1",
-        freeze_time: "2025-03-20 16:49:38",
-        temp_probe_1: -27.543,
-        row_in_csv: 5965,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A10",
-        freeze_time: "2025-03-20 16:35:04",
-        temp_probe_1: -22.863,
-        row_in_csv: 4131,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A11",
-        freeze_time: "2025-03-20 16:32:47",
-        temp_probe_1: -22.151,
-        row_in_csv: 3994,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A12",
-        freeze_time: "2025-03-20 16:34:09",
-        temp_probe_1: -22.588,
-        row_in_csv: 4076,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A2",
-        freeze_time: "2025-03-20 16:48:49",
-        temp_probe_1: -27.278,
-        row_in_csv: 5936,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A3",
-        freeze_time: "2025-03-20 16:50:39",
-        temp_probe_1: -27.856,
-        row_in_csv: 6001,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A4",
-        freeze_time: "2025-03-20 16:50:58",
-        temp_probe_1: -27.955,
-        row_in_csv: 6012,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A5",
-        freeze_time: "2025-03-20 16:45:19",
-        temp_probe_1: -26.121,
-        row_in_csv: 5726,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A6",
-        freeze_time: "2025-03-20 16:51:07",
-        temp_probe_1: -28.003,
-        row_in_csv: 6017,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A7",
-        freeze_time: "2025-03-20 16:33:57",
-        temp_probe_1: -22.522,
-        row_in_csv: 4069,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A8",
-        freeze_time: "2025-03-20 16:34:41",
-        temp_probe_1: -22.753,
-        row_in_csv: 4095,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "A9",
-        freeze_time: "2025-03-20 16:42:06",
-        temp_probe_1: -25.078,
-        row_in_csv: 5533,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B1",
-        freeze_time: "2025-03-20 16:42:35",
-        temp_probe_1: -25.232,
-        row_in_csv: 5550,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B10",
-        freeze_time: "2025-03-20 16:42:44",
-        temp_probe_1: -25.278,
-        row_in_csv: 5555,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B11",
-        freeze_time: "2025-03-20 16:43:21",
-        temp_probe_1: -25.469,
-        row_in_csv: 5577,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B12",
-        freeze_time: "2025-03-20 16:33:20",
-        temp_probe_1: -22.322,
-        row_in_csv: 4047,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B2",
-        freeze_time: "2025-03-20 16:46:10",
-        temp_probe_1: -26.419,
-        row_in_csv: 5756,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B3",
-        freeze_time: "2025-03-20 16:52:23",
-        temp_probe_1: -28.435,
-        row_in_csv: 6062,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B4",
-        freeze_time: "2025-03-20 16:46:09",
-        temp_probe_1: -26.412,
-        row_in_csv: 5755,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B5",
-        freeze_time: "2025-03-20 16:40:09",
-        temp_probe_1: -24.460,
-        row_in_csv: 5463,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B6",
-        freeze_time: "2025-03-20 16:46:35",
-        temp_probe_1: -26.555,
-        row_in_csv: 5770,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B7",
-        freeze_time: "2025-03-20 16:46:33",
-        temp_probe_1: -26.550,
-        row_in_csv: 5769,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B8",
-        freeze_time: "2025-03-20 16:46:30",
-        temp_probe_1: -26.531,
-        row_in_csv: 5767,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "B9",
-        freeze_time: "2025-03-20 16:42:04",
-        temp_probe_1: -25.065,
-        row_in_csv: 5532,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C1",
-        freeze_time: "2025-03-20 16:50:22",
-        temp_probe_1: -27.769,
-        row_in_csv: 5991,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C10",
-        freeze_time: "2025-03-20 16:31:47",
-        temp_probe_1: -21.823,
-        row_in_csv: 3971,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C11",
-        freeze_time: "2025-03-20 16:36:24",
-        temp_probe_1: -23.257,
-        row_in_csv: 4161,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C12",
-        freeze_time: "2025-03-20 16:36:50",
-        temp_probe_1: -23.392,
-        row_in_csv: 4176,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C2",
-        freeze_time: "2025-03-20 16:46:24",
-        temp_probe_1: -26.500,
-        row_in_csv: 5763,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C3",
-        freeze_time: "2025-03-20 16:49:16",
-        temp_probe_1: -27.425,
-        row_in_csv: 5952,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C4",
-        freeze_time: "2025-03-20 16:47:00",
-        temp_probe_1: -26.678,
-        row_in_csv: 5784,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C5",
-        freeze_time: "2025-03-20 16:40:48",
-        temp_probe_1: -24.659,
-        row_in_csv: 5486,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C6",
-        freeze_time: "2025-03-20 16:36:39",
-        temp_probe_1: -23.335,
-        row_in_csv: 4170,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C7",
-        freeze_time: "2025-03-20 16:42:59",
-        temp_probe_1: -25.358,
-        row_in_csv: 5564,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C8",
-        freeze_time: "2025-03-20 16:48:32",
-        temp_probe_1: -27.181,
-        row_in_csv: 5926,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "C9",
-        freeze_time: "2025-03-20 16:43:37",
-        temp_probe_1: -25.549,
-        row_in_csv: 5587,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D1",
-        freeze_time: "2025-03-20 16:41:27",
-        temp_probe_1: -24.870,
-        row_in_csv: 5509,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D10",
-        freeze_time: "2025-03-20 16:35:40",
-        temp_probe_1: -23.047,
-        row_in_csv: 4135,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D11",
-        freeze_time: "2025-03-20 16:29:30",
-        temp_probe_1: -21.084,
-        row_in_csv: 3858,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D12",
-        freeze_time: "2025-03-20 16:37:05",
-        temp_probe_1: -23.469,
-        row_in_csv: 4185,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D2",
-        freeze_time: "2025-03-20 16:41:26",
-        temp_probe_1: -24.864,
-        row_in_csv: 5508,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D3",
-        freeze_time: "2025-03-20 16:49:17",
-        temp_probe_1: -27.430,
-        row_in_csv: 5953,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D4",
-        freeze_time: "2025-03-20 16:43:02",
-        temp_probe_1: -25.377,
-        row_in_csv: 5566,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D5",
-        freeze_time: "2025-03-20 16:45:53",
-        temp_probe_1: -26.322,
-        row_in_csv: 5746,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D6",
-        freeze_time: "2025-03-20 16:49:27",
-        temp_probe_1: -27.487,
-        row_in_csv: 5959,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D7",
-        freeze_time: "2025-03-20 16:42:35",
-        temp_probe_1: -25.232,
-        row_in_csv: 5550,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D8",
-        freeze_time: "2025-03-20 16:38:05",
-        temp_probe_1: -23.793,
-        row_in_csv: 4219,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "D9",
-        freeze_time: "2025-03-20 16:53:43",
-        temp_probe_1: -28.887,
-        row_in_csv: 6109,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E1",
-        freeze_time: "2025-03-20 16:47:31",
-        temp_probe_1: -26.832,
-        row_in_csv: 5802,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E10",
-        freeze_time: "2025-03-20 16:42:16",
-        temp_probe_1: -25.132,
-        row_in_csv: 5538,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E11",
-        freeze_time: "2025-03-20 16:39:50",
-        temp_probe_1: -24.360,
-        row_in_csv: 5452,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E12",
-        freeze_time: "2025-03-20 16:35:19",
-        temp_probe_1: -22.943,
-        row_in_csv: 4123,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E2",
-        freeze_time: "2025-03-20 16:53:24",
-        temp_probe_1: -28.786,
-        row_in_csv: 6098,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E3",
-        freeze_time: "2025-03-20 16:53:37",
-        temp_probe_1: -28.858,
-        row_in_csv: 6106,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E4",
-        freeze_time: "2025-03-20 16:42:33",
-        temp_probe_1: -25.221,
-        row_in_csv: 5549,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E5",
-        freeze_time: "2025-03-20 16:42:09",
-        temp_probe_1: -25.095,
-        row_in_csv: 5535,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E6",
-        freeze_time: "2025-03-20 16:43:46",
-        temp_probe_1: -25.592,
-        row_in_csv: 5593,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E7",
-        freeze_time: "2025-03-20 16:46:40",
-        temp_probe_1: -26.581,
-        row_in_csv: 5773,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E8",
-        freeze_time: "2025-03-20 16:35:30",
-        temp_probe_1: -22.998,
-        row_in_csv: 4129,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "E9",
-        freeze_time: "2025-03-20 16:50:38",
-        temp_probe_1: -27.850,
-        row_in_csv: 6000,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F1",
-        freeze_time: "2025-03-20 16:46:48",
-        temp_probe_1: -26.619,
-        row_in_csv: 5778,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F10",
-        freeze_time: "2025-03-20 16:39:14",
-        temp_probe_1: -24.180,
-        row_in_csv: 5431,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F11",
-        freeze_time: "2025-03-20 16:37:17",
-        temp_probe_1: -23.539,
-        row_in_csv: 4192,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F12",
-        freeze_time: "2025-03-20 16:28:21",
-        temp_probe_1: -20.719,
-        row_in_csv: 3817,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F2",
-        freeze_time: "2025-03-20 16:49:33",
-        temp_probe_1: -27.518,
-        row_in_csv: 5962,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F3",
-        freeze_time: "2025-03-20 16:43:39",
-        temp_probe_1: -25.560,
-        row_in_csv: 5588,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F4",
-        freeze_time: "2025-03-20 16:50:29",
-        temp_probe_1: -27.805,
-        row_in_csv: 5995,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F5",
-        freeze_time: "2025-03-20 16:52:05",
-        temp_probe_1: -28.330,
-        row_in_csv: 6051,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F6",
-        freeze_time: "2025-03-20 16:48:24",
-        temp_probe_1: -27.133,
-        row_in_csv: 5921,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F7",
-        freeze_time: "2025-03-20 16:36:28",
-        temp_probe_1: -23.277,
-        row_in_csv: 4164,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F8",
-        freeze_time: "2025-03-20 16:39:34",
-        temp_probe_1: -24.287,
-        row_in_csv: 5443,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "F9",
-        freeze_time: "2025-03-20 16:48:15",
-        temp_probe_1: -27.088,
-        row_in_csv: 5916,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G1",
-        freeze_time: "2025-03-20 16:40:36",
-        temp_probe_1: -24.597,
-        row_in_csv: 5479,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G10",
-        freeze_time: "2025-03-20 16:28:37",
-        temp_probe_1: -20.803,
-        row_in_csv: 3826,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G11",
-        freeze_time: "2025-03-20 16:40:16",
-        temp_probe_1: -24.493,
-        row_in_csv: 5467,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G12",
-        freeze_time: "2025-03-20 16:34:46",
-        temp_probe_1: -22.777,
-        row_in_csv: 4098,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G2",
-        freeze_time: "2025-03-20 16:43:23",
-        temp_probe_1: -25.481,
-        row_in_csv: 5578,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G3",
-        freeze_time: "2025-03-20 16:49:36",
-        temp_probe_1: -27.535,
-        row_in_csv: 5973,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G4",
-        freeze_time: "2025-03-20 16:53:04",
-        temp_probe_1: -28.673,
-        row_in_csv: 6086,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G5",
-        freeze_time: "2025-03-20 16:41:01",
-        temp_probe_1: -24.728,
-        row_in_csv: 5494,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G6",
-        freeze_time: "2025-03-20 16:50:39",
-        temp_probe_1: -27.856,
-        row_in_csv: 6001,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G7",
-        freeze_time: "2025-03-20 16:41:07",
-        temp_probe_1: -24.760,
-        row_in_csv: 5497,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G8",
-        freeze_time: "2025-03-20 16:43:10",
-        temp_probe_1: -25.416,
-        row_in_csv: 5571,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "G9",
-        freeze_time: "2025-03-20 16:39:38",
-        temp_probe_1: -24.308,
-        row_in_csv: 5445,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H1",
-        freeze_time: "2025-03-20 16:43:58",
-        temp_probe_1: -25.653,
-        row_in_csv: 5600,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H10",
-        freeze_time: "2025-03-20 16:37:19",
-        temp_probe_1: -23.551,
-        row_in_csv: 4193,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H11",
-        freeze_time: "2025-03-20 16:37:45",
-        temp_probe_1: -23.693,
-        row_in_csv: 4208,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H12",
-        freeze_time: "2025-03-20 16:38:58",
-        temp_probe_1: -24.087,
-        row_in_csv: 4251,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H2",
-        freeze_time: "2025-03-20 16:48:34",
-        temp_probe_1: -27.192,
-        row_in_csv: 5927,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H3",
-        freeze_time: "2025-03-20 16:52:49",
-        temp_probe_1: -28.589,
-        row_in_csv: 6077,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H4",
-        freeze_time: "2025-03-20 16:43:06",
-        temp_probe_1: -25.396,
-        row_in_csv: 5569,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H5",
-        freeze_time: "2025-03-20 16:46:49",
-        temp_probe_1: -26.624,
-        row_in_csv: 5779,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H6",
-        freeze_time: "2025-03-20 16:43:16",
-        temp_probe_1: -25.444,
-        row_in_csv: 5575,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H7",
-        freeze_time: "2025-03-20 16:38:58",
-        temp_probe_1: -24.087,
-        row_in_csv: 4251,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H8",
-        freeze_time: "2025-03-20 16:44:22",
-        temp_probe_1: -25.784,
-        row_in_csv: 5611,
-    },
-    WellTransitionData {
-        tray: "P1",
-        coordinate: "H9",
-        freeze_time: "2025-03-20 16:50:08",
-        temp_probe_1: -27.704,
-        row_in_csv: 5983,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A1",
-        freeze_time: "2025-03-20 16:33:53",
-        temp_probe_1: -22.502,
-        row_in_csv: 4066,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A10",
-        freeze_time: "2025-03-20 16:43:28",
-        temp_probe_1: -25.504,
-        row_in_csv: 5581,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A11",
-        freeze_time: "2025-03-20 16:40:28",
-        temp_probe_1: -24.555,
-        row_in_csv: 5474,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A12",
-        freeze_time: "2025-03-20 16:41:13",
-        temp_probe_1: -24.792,
-        row_in_csv: 5500,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A2",
-        freeze_time: "2025-03-20 16:31:55",
-        temp_probe_1: -21.867,
-        row_in_csv: 3975,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A3",
-        freeze_time: "2025-03-20 16:49:22",
-        temp_probe_1: -27.458,
-        row_in_csv: 5956,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A4",
-        freeze_time: "2025-03-20 16:45:52",
-        temp_probe_1: -26.314,
-        row_in_csv: 5745,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A5",
-        freeze_time: "2025-03-20 16:48:10",
-        temp_probe_1: -27.059,
-        row_in_csv: 5914,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A6",
-        freeze_time: "2025-03-20 16:43:18",
-        temp_probe_1: -25.455,
-        row_in_csv: 5576,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A7",
-        freeze_time: "2025-03-20 16:45:19",
-        temp_probe_1: -26.121,
-        row_in_csv: 5726,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A8",
-        freeze_time: "2025-03-20 16:42:00",
-        temp_probe_1: -25.043,
-        row_in_csv: 5530,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "A9",
-        freeze_time: "2025-03-20 16:35:39",
-        temp_probe_1: -23.045,
-        row_in_csv: 4134,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B1",
-        freeze_time: "2025-03-20 16:38:42",
-        temp_probe_1: -24.000,
-        row_in_csv: 4242,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B10",
-        freeze_time: "2025-03-20 16:43:44",
-        temp_probe_1: -25.580,
-        row_in_csv: 5590,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B11",
-        freeze_time: "2025-03-20 16:37:47",
-        temp_probe_1: -23.704,
-        row_in_csv: 4209,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B12",
-        freeze_time: "2025-03-20 16:42:39",
-        temp_probe_1: -25.252,
-        row_in_csv: 5553,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B2",
-        freeze_time: "2025-03-20 16:36:15",
-        temp_probe_1: -23.217,
-        row_in_csv: 4154,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B3",
-        freeze_time: "2025-03-20 16:50:10",
-        temp_probe_1: -27.709,
-        row_in_csv: 5984,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B4",
-        freeze_time: "2025-03-20 16:35:47",
-        temp_probe_1: -23.083,
-        row_in_csv: 4139,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B5",
-        freeze_time: "2025-03-20 16:50:28",
-        temp_probe_1: -27.799,
-        row_in_csv: 5994,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B6",
-        freeze_time: "2025-03-20 16:50:10",
-        temp_probe_1: -27.709,
-        row_in_csv: 5984,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B7",
-        freeze_time: "2025-03-20 16:50:12",
-        temp_probe_1: -27.718,
-        row_in_csv: 5985,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B8",
-        freeze_time: "2025-03-20 16:19:47",
-        temp_probe_1: -17.969,
-        row_in_csv: 3016,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "B9",
-        freeze_time: "2025-03-20 16:42:47",
-        temp_probe_1: -25.295,
-        row_in_csv: 5558,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C1",
-        freeze_time: "2025-03-20 16:41:03",
-        temp_probe_1: -24.738,
-        row_in_csv: 5495,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C10",
-        freeze_time: "2025-03-20 16:31:52",
-        temp_probe_1: -21.852,
-        row_in_csv: 3973,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C11",
-        freeze_time: "2025-03-20 16:41:46",
-        temp_probe_1: -24.966,
-        row_in_csv: 5520,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C12",
-        freeze_time: "2025-03-20 16:35:50",
-        temp_probe_1: -23.094,
-        row_in_csv: 4141,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C2",
-        freeze_time: "2025-03-20 16:38:52",
-        temp_probe_1: -24.056,
-        row_in_csv: 4248,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C3",
-        freeze_time: "2025-03-20 16:34:38",
-        temp_probe_1: -22.735,
-        row_in_csv: 4093,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C4",
-        freeze_time: "2025-03-20 16:39:22",
-        temp_probe_1: -24.221,
-        row_in_csv: 5434,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C5",
-        freeze_time: "2025-03-20 16:52:23",
-        temp_probe_1: -28.435,
-        row_in_csv: 6062,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C6",
-        freeze_time: "2025-03-20 16:40:53",
-        temp_probe_1: -24.685,
-        row_in_csv: 5489,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C7",
-        freeze_time: "2025-03-20 16:36:56",
-        temp_probe_1: -23.424,
-        row_in_csv: 4180,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C8",
-        freeze_time: "2025-03-20 16:43:04",
-        temp_probe_1: -25.388,
-        row_in_csv: 5567,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "C9",
-        freeze_time: "2025-03-20 16:43:59",
-        temp_probe_1: -25.662,
-        row_in_csv: 5601,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D1",
-        freeze_time: "2025-03-20 16:36:23",
-        temp_probe_1: -23.253,
-        row_in_csv: 4160,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D10",
-        freeze_time: "2025-03-20 16:42:08",
-        temp_probe_1: -25.088,
-        row_in_csv: 5534,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D11",
-        freeze_time: "2025-03-20 16:42:54",
-        temp_probe_1: -25.334,
-        row_in_csv: 5561,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D12",
-        freeze_time: "2025-03-20 16:39:36",
-        temp_probe_1: -24.295,
-        row_in_csv: 5442,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D2",
-        freeze_time: "2025-03-20 16:28:12",
-        temp_probe_1: -20.669,
-        row_in_csv: 3812,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D3",
-        freeze_time: "2025-03-20 16:37:51",
-        temp_probe_1: -23.724,
-        row_in_csv: 4211,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D4",
-        freeze_time: "2025-03-20 16:54:47",
-        temp_probe_1: -29.247,
-        row_in_csv: 6147,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D5",
-        freeze_time: "2025-03-20 16:48:54",
-        temp_probe_1: -27.303,
-        row_in_csv: 5939,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D6",
-        freeze_time: "2025-03-20 16:51:20",
-        temp_probe_1: -28.076,
-        row_in_csv: 6025,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D7",
-        freeze_time: "2025-03-20 16:46:36",
-        temp_probe_1: -26.564,
-        row_in_csv: 5771,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D8",
-        freeze_time: "2025-03-20 16:38:39",
-        temp_probe_1: -23.983,
-        row_in_csv: 4241,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "D9",
-        freeze_time: "2025-03-20 16:38:50",
-        temp_probe_1: -24.045,
-        row_in_csv: 4247,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E1",
-        freeze_time: "2025-03-20 16:36:06",
-        temp_probe_1: -23.174,
-        row_in_csv: 4150,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E10",
-        freeze_time: "2025-03-20 16:42:15",
-        temp_probe_1: -25.126,
-        row_in_csv: 5537,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E11",
-        freeze_time: "2025-03-20 16:41:46",
-        temp_probe_1: -24.966,
-        row_in_csv: 5520,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E12",
-        freeze_time: "2025-03-20 16:40:11",
-        temp_probe_1: -24.468,
-        row_in_csv: 5464,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E2",
-        freeze_time: "2025-03-20 16:34:37",
-        temp_probe_1: -22.733,
-        row_in_csv: 4092,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E3",
-        freeze_time: "2025-03-20 16:45:28",
-        temp_probe_1: -26.176,
-        row_in_csv: 5731,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E4",
-        freeze_time: "2025-03-20 16:45:03",
-        temp_probe_1: -26.022,
-        row_in_csv: 5716,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E5",
-        freeze_time: "2025-03-20 16:44:14",
-        temp_probe_1: -25.740,
-        row_in_csv: 5687,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E6",
-        freeze_time: "2025-03-20 16:43:03",
-        temp_probe_1: -25.382,
-        row_in_csv: 5565,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E7",
-        freeze_time: "2025-03-20 16:56:25",
-        temp_probe_1: -29.793,
-        row_in_csv: 6205,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E8",
-        freeze_time: "2025-03-20 16:41:17",
-        temp_probe_1: -24.814,
-        row_in_csv: 5503,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "E9",
-        freeze_time: "2025-03-20 16:26:37",
-        temp_probe_1: -20.169,
-        row_in_csv: 3753,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F1",
-        freeze_time: "2025-03-20 16:34:15",
-        temp_probe_1: -22.623,
-        row_in_csv: 4079,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F10",
-        freeze_time: "2025-03-20 16:39:23",
-        temp_probe_1: -24.224,
-        row_in_csv: 5435,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F11",
-        freeze_time: "2025-03-20 16:40:27",
-        temp_probe_1: -24.548,
-        row_in_csv: 5473,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F12",
-        freeze_time: "2025-03-20 16:39:58",
-        temp_probe_1: -24.401,
-        row_in_csv: 5456,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F2",
-        freeze_time: "2025-03-20 16:23:35",
-        temp_probe_1: -19.192,
-        row_in_csv: 3572,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F3",
-        freeze_time: "2025-03-20 16:48:02",
-        temp_probe_1: -27.009,
-        row_in_csv: 5909,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F4",
-        freeze_time: "2025-03-20 16:45:24",
-        temp_probe_1: -26.152,
-        row_in_csv: 5729,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F5",
-        freeze_time: "2025-03-20 16:51:34",
-        temp_probe_1: -28.150,
-        row_in_csv: 6033,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F6",
-        freeze_time: "2025-03-20 16:42:18",
-        temp_probe_1: -25.139,
-        row_in_csv: 5539,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F7",
-        freeze_time: "2025-03-20 16:47:16",
-        temp_probe_1: -26.755,
-        row_in_csv: 5793,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F8",
-        freeze_time: "2025-03-20 16:29:55",
-        temp_probe_1: -21.217,
-        row_in_csv: 3873,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "F9",
-        freeze_time: "2025-03-20 16:43:44",
-        temp_probe_1: -25.580,
-        row_in_csv: 5590,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G1",
-        freeze_time: "2025-03-20 16:37:20",
-        temp_probe_1: -23.553,
-        row_in_csv: 4194,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G10",
-        freeze_time: "2025-03-20 16:36:03",
-        temp_probe_1: -23.158,
-        row_in_csv: 4148,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G11",
-        freeze_time: "2025-03-20 16:41:01",
-        temp_probe_1: -24.728,
-        row_in_csv: 5494,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G12",
-        freeze_time: "2025-03-20 16:37:46",
-        temp_probe_1: -23.698,
-        row_in_csv: 4209,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G2",
-        freeze_time: "2025-03-20 16:35:45",
-        temp_probe_1: -23.071,
-        row_in_csv: 4137,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G3",
-        freeze_time: "2025-03-20 16:41:00",
-        temp_probe_1: -24.723,
-        row_in_csv: 5493,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G4",
-        freeze_time: "2025-03-20 16:44:12",
-        temp_probe_1: -25.729,
-        row_in_csv: 5686,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G5",
-        freeze_time: "2025-03-20 16:46:44",
-        temp_probe_1: -26.599,
-        row_in_csv: 5775,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G6",
-        freeze_time: "2025-03-20 16:40:40",
-        temp_probe_1: -24.613,
-        row_in_csv: 5481,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G7",
-        freeze_time: "2025-03-20 16:46:29",
-        temp_probe_1: -26.521,
-        row_in_csv: 5766,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G8",
-        freeze_time: "2025-03-20 16:39:16",
-        temp_probe_1: -24.190,
-        row_in_csv: 5432,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "G9",
-        freeze_time: "2025-03-20 16:44:42",
-        temp_probe_1: -25.898,
-        row_in_csv: 5703,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H1",
-        freeze_time: "2025-03-20 16:36:29",
-        temp_probe_1: -23.284,
-        row_in_csv: 4165,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H10",
-        freeze_time: "2025-03-20 16:44:30",
-        temp_probe_1: -25.832,
-        row_in_csv: 5696,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H11",
-        freeze_time: "2025-03-20 16:34:40",
-        temp_probe_1: -22.746,
-        row_in_csv: 4094,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H12",
-        freeze_time: "2025-03-20 16:35:40",
-        temp_probe_1: -23.047,
-        row_in_csv: 4135,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H2",
-        freeze_time: "2025-03-20 16:35:30",
-        temp_probe_1: -22.998,
-        row_in_csv: 4129,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H3",
-        freeze_time: "2025-03-20 16:46:20",
-        temp_probe_1: -26.477,
-        row_in_csv: 5758,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H4",
-        freeze_time: "2025-03-20 16:46:46",
-        temp_probe_1: -26.608,
-        row_in_csv: 5774,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H5",
-        freeze_time: "2025-03-20 16:49:20",
-        temp_probe_1: -27.449,
-        row_in_csv: 5955,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H6",
-        freeze_time: "2025-03-20 16:38:08",
-        temp_probe_1: -23.808,
-        row_in_csv: 4221,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H7",
-        freeze_time: "2025-03-20 16:50:11",
-        temp_probe_1: -27.712,
-        row_in_csv: 5985,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H8",
-        freeze_time: "2025-03-20 16:39:00",
-        temp_probe_1: -24.101,
-        row_in_csv: 4252,
-    },
-    WellTransitionData {
-        tray: "P2",
-        coordinate: "H9",
-        freeze_time: "2025-03-20 16:39:21",
-        temp_probe_1: -24.215,
-        row_in_csv: 4264,
-    },
-];
-
-/// Test constants derived from merged.csv
-pub const EXPECTED_EXPERIMENT_START: &str = "2025-03-20 15:13:47";
-pub const EXPECTED_FIRST_FREEZE: &str = "2025-03-20 16:19:47";
-pub const EXPECTED_TOTAL_TIME_POINTS: usize = 6786;
-pub const EXPECTED_TOTAL_WELLS: usize = 192;
-pub const EXPECTED_TEMPERATURE_PROBES: usize = 8;
-
-/// Create a tray configuration with embedded trays (post-flattening structure)
-async fn create_test_tray_config_with_trays(app: &Router, name: &str) -> String {
-    let tray_config_data = json!({
-        "name": name,
-        "experiment_default": false,
-        "trays": [
-            {
-                "order_sequence": 1,
-                "rotation_degrees": 0,
-                "name": "P1",
-                "qty_x_axis": 8,
-                "qty_y_axis": 12,
-                "well_relative_diameter": 2.5
-            },
-            {
-                "order_sequence": 2,
-                "rotation_degrees": 0,
-                "name": "P2",
-                "qty_x_axis": 8,
-                "qty_y_axis": 12,
-                "well_relative_diameter": 2.5
-            }
-        ]
-    });
-
-    println!(
-        "🏗️ Creating tray configuration '{}' with embedded P1/P2 trays: {}",
-        name, tray_config_data
-    );
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/trays")
-                .header("content-type", "application/json")
-                .body(Body::from(tray_config_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let status = response.status();
-    let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let body_str = String::from_utf8(body.to_vec()).unwrap();
-
-    if status != StatusCode::CREATED {
-        println!("❌ Failed to create tray config");
-        println!("   Status: {}", status);
-        println!("   Request payload: {}", tray_config_data);
-        println!("   Response body: {}", body_str);
-
-        // Try to parse the error message from JSON
-        if let Ok(error_json) = serde_json::from_str::<serde_json::Value>(&body_str) {
-            println!("   Parsed error: {:?}", error_json);
-        }
-
-        panic!(
-            "Failed to create tray config. Status: {}, Body: {}",
-            status, body_str
-        );
-    }
-
-    body_str
-}
-
-/// Upload Excel file via API with proper multipart support
-async fn upload_excel_file(app: &Router, experiment_id: &str) -> Value {
-    // Read the test Excel file
-    let excel_data = fs::read("/home/evan/projects/EERL/SPICE/spice-api/src/routes/experiments/test_resources/merged.xlsx")
-        .expect("Should find test Excel file");
-
-    // Create a properly formatted multipart body with correct boundaries and headers
-    let boundary = "----formdata-test-boundary-123456789";
-    let mut body = Vec::new();
-
-    // Construct multipart body according to RFC 7578
-    body.extend_from_slice(format!("--{}\r\n", boundary).as_bytes());
-    body.extend_from_slice(
-        b"Content-Disposition: form-data; name=\"file\"; filename=\"merged.xlsx\"\r\n",
-    );
-    body.extend_from_slice(
-        b"Content-Type: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet\r\n",
-    );
-    body.extend_from_slice(b"\r\n");
-    body.extend_from_slice(&excel_data);
-    body.extend_from_slice(b"\r\n");
-    body.extend_from_slice(format!("--{}--\r\n", boundary).as_bytes());
-
-    println!("   📤 Multipart body size: {} bytes", body.len());
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri(format!("/api/experiments/{experiment_id}/process-excel"))
-                .header(
-                    "content-type",
-                    format!("multipart/form-data; boundary={}", boundary),
-                )
-                .body(Body::from(body))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let status_code = response.status();
-    let response_body = axum::body::to_bytes(response.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let body_str = String::from_utf8(response_body.to_vec()).unwrap();
-
-    json!({
-        "status_code": status_code.as_u16(),
-        "body": body_str
-    })
-}
-
-#[tokio::test]
-async fn test_comprehensive_excel_validation_with_specific_transitions() {
-    let app = setup_test_app().await;
-
-    println!("🔬 Starting comprehensive Excel validation test...");
-
-    // Step 1: Create experiment with proper tray configuration
-    let tray_config_response =
-        create_test_tray_config_with_trays(&app, "Comprehensive Test Config").await;
-    let tray_config: Value = serde_json::from_str(&tray_config_response).unwrap();
-
-    let tray_config_id = tray_config["id"].as_str().unwrap();
-
-    let experiment_payload = serde_json::json!({
-        "name": "Comprehensive Validation Test",
-        "remarks": "Testing specific well transitions from merged.csv",
-        "tray_configuration_id": tray_config_id,
-        "is_calibration": false
-    });
-
-    let experiment_response = app
-        .clone()
-        .oneshot(
-            axum::http::Request::builder()
-                .method(axum::http::Method::POST)
-                .uri("/api/experiments")
-                .header("content-type", "application/json")
-                .body(axum::body::Body::from(experiment_payload.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let exp_status = experiment_response.status();
-    let exp_body = axum::body::to_bytes(experiment_response.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let exp_body_str = String::from_utf8(exp_body.to_vec()).unwrap();
-
-    if exp_status != StatusCode::OK && exp_status != StatusCode::CREATED {
-        println!("❌ Failed to create experiment");
-        println!("   Status: {}", exp_status);
-        println!("   Request payload: {}", experiment_payload);
-        println!("   Response body: {}", exp_body_str);
-    }
-
-    assert_eq!(exp_status, 201);
-    let experiment: Value = serde_json::from_str(&exp_body_str).unwrap();
-    let experiment_id = experiment["id"].as_str().unwrap();
-
-    println!("✅ Created experiment: {}", experiment_id);
-
-    // Step 2: Upload Excel file and process
-    let upload_result = upload_excel_file(&app, experiment_id).await;
-    println!("📤 Excel upload result: {:?}", upload_result);
-
-    assert!(
-        upload_result["body"]
-            .as_str()
-            .unwrap()
-            .contains("completed")
-    );
-
-    // Step 3: Fetch experiment results with comprehensive validation
-    let results_response = app
-        .clone()
-        .oneshot(
-            axum::http::Request::builder()
-                .method(axum::http::Method::GET)
-                .uri(&format!("/api/experiments/{}", experiment_id))
-                .body(axum::body::Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(results_response.status(), 200);
-    let results_body = axum::body::to_bytes(results_response.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let experiment_with_results: Value = serde_json::from_slice(&results_body).unwrap();
-    let results_summary = &experiment_with_results["results_summary"];
-
-    // Step 4: Validate high-level counts
-    validate_experiment_totals(results_summary);
-
-    // Step 5: Validate specific well transitions
-    validate_specific_well_transitions(&experiment_with_results);
-
-    // Step 6: Validate temperature data accuracy
-    validate_temperature_readings(&experiment_with_results);
-
-    // Step 7: Validate timing accuracy
-    validate_experiment_timing(results_summary);
-
-    println!("🎉 All comprehensive validations passed!");
-}
-
-fn validate_experiment_totals(results_summary: &Value) {
-    println!("🔢 Validating experiment totals...");
-
-    let total_wells = results_summary["total_wells"].as_u64().unwrap_or(0);
-    let wells_with_data = results_summary["wells_with_data"].as_u64().unwrap_or(0);
-    let wells_frozen = results_summary["wells_frozen"].as_u64().unwrap_or(0);
-    let total_time_points = results_summary["total_time_points"].as_u64().unwrap_or(0);
-
-    assert_eq!(
-        total_wells, EXPECTED_TOTAL_WELLS as u64,
-        "Total wells should be {}, got {}",
-        EXPECTED_TOTAL_WELLS, total_wells
-    );
-    assert_eq!(
-        wells_with_data, EXPECTED_TOTAL_WELLS as u64,
-        "Wells with data should be {}, got {}",
-        EXPECTED_TOTAL_WELLS, wells_with_data
-    );
-    assert_eq!(
-        wells_frozen, EXPECTED_TOTAL_WELLS as u64,
-        "All wells should be frozen, got {}",
-        wells_frozen
-    );
-    assert_eq!(
-        total_time_points, EXPECTED_TOTAL_TIME_POINTS as u64,
-        "Time points should be {}, got {}",
-        EXPECTED_TOTAL_TIME_POINTS, total_time_points
-    );
-
-    println!("   ✅ Total wells: {} ✓", total_wells);
-    println!("   ✅ Wells with data: {} ✓", wells_with_data);
-    println!("   ✅ Wells frozen: {} ✓", wells_frozen);
-    println!("   ✅ Time points: {} ✓", total_time_points);
-}
-
-fn validate_specific_well_transitions(experiment: &Value) {
-    println!("🎯 Validating specific well transitions...");
-
-    let well_summaries = experiment["results_summary"]["well_summaries"]
-        .as_array()
-        .expect("Should have well summaries");
-
-    // Create lookup map by tray and coordinate
-    let mut well_lookup: HashMap<String, &Value> = HashMap::new();
-    for well in well_summaries {
-        let tray_name = well["tray_name"].as_str().unwrap_or("unknown");
-        let coordinate = well["coordinate"].as_str().unwrap_or("unknown");
-        let key = format!("{}_{}", tray_name, coordinate);
-        well_lookup.insert(key, well);
-    }
-
-    println!("   📋 Created lookup for {} wells", well_lookup.len());
-
-    // Validate each expected transition
-    for expected in EXPECTED_TRANSITIONS {
-        let key = format!("{}_{}", expected.tray, expected.coordinate);
-        let well = well_lookup
-            .get(&key)
-            .unwrap_or_else(|| panic!("Could not find well {}", key));
-
-        // Validate well has a freeze time
-        let freeze_time = well["first_phase_change_time"]
-            .as_str()
-            .unwrap_or_else(|| panic!("Well {} should have first_phase_change_time", key));
-
-        // Validate final state is frozen
-        let final_state = well["final_state"].as_str().unwrap_or("unknown");
-        assert_eq!(final_state, "frozen", "Well {} should be frozen", key);
-
-        // Validate temperature probes exist
-        let temp_probes = &well["first_phase_change_temperature_probes"];
-        assert!(
-            temp_probes.is_object(),
-            "Well {} should have temperature probe data",
-            key
-        );
-
-        // Temperature values are stored as strings (Decimal), need to parse them
-        let probe1_temp = temp_probes["probe_1"]
-            .as_str()
-            .and_then(|s| s.parse::<f64>().ok())
-            .unwrap_or_else(|| panic!("Well {} should have probe_1 temperature", key));
-
-        // Allow tolerance for difference between averaged (CSV analysis) and single probe (API) temperatures
-        // Since CSV analysis used 8-probe averages but API uses individual probe readings
-        let temp_diff = (probe1_temp - expected.temp_probe_1).abs();
-        assert!(
-            temp_diff < 1.0,
-            "Well {} probe 1 temperature should be ~{}°C, got {}°C (diff: {}°C)",
-            key,
-            expected.temp_probe_1,
-            probe1_temp,
-            temp_diff
-        );
-
-        println!(
-            "   ✅ Well {}: froze at {}, temp={}°C ✓",
-            key, freeze_time, probe1_temp
-        );
-    }
-
-    println!(
-        "   🎯 Validated {} specific transitions",
-        EXPECTED_TRANSITIONS.len()
-    );
-
-    // Report validation coverage
-    println!("   📊 Validation Coverage:");
-    println!(
-        "      🔸 Total wells validated: {}/192 ({:.1}%)",
-        EXPECTED_TRANSITIONS.len(),
-        (EXPECTED_TRANSITIONS.len() as f64 / 192.0) * 100.0
-    );
-
-    if EXPECTED_TRANSITIONS.len() == 192 {
-        println!("      🎉 COMPLETE COVERAGE: All 192 wells validated!");
-    }
-}
-
-fn validate_temperature_readings(_experiment: &Value) {
-    println!("🌡️  Validating temperature readings...");
-
-    // Temperature validation would require time series data
-    // For now, validate that temperature probe structure exists
-    println!("   ✅ Temperature probe structure validated");
-}
-
-fn validate_experiment_timing(results_summary: &Value) {
-    println!("⏰ Validating experiment timing...");
-
-    let first_timestamp = results_summary["first_timestamp"]
-        .as_str()
-        .expect("Should have first_timestamp");
-    let last_timestamp = results_summary["last_timestamp"]
-        .as_str()
-        .expect("Should have last_timestamp");
-
-    // Validate experiment start time matches expected
-    assert!(
-        first_timestamp.contains("2025-03-20"),
-        "Experiment should start on 2025-03-20, got {}",
-        first_timestamp
-    );
-    assert!(
-        first_timestamp.contains("15:13"),
-        "Experiment should start around 15:13, got {}",
-        first_timestamp
-    );
-
-    println!("   ✅ Experiment start: {} ✓", first_timestamp);
-    println!("   ✅ Experiment end: {} ✓", last_timestamp);
-
-    // Calculate duration (should be about 1 hour 6 minutes based on CSV)
-    // This is a rough validation - exact timing depends on processing
-    println!("   ✅ Timing validation complete");
-}
-
-#[tokio::test]
-async fn test_well_coordinate_mapping_accuracy() {
-    println!("🗺️  Testing well coordinate mapping accuracy...");
-
-    let app = setup_test_app().await;
-
-    // Create experiment and upload
-    let tray_config_response = create_test_tray_config_with_trays(&app, "Coordinate Test").await;
-    let tray_config: Value = serde_json::from_str(&tray_config_response).unwrap();
-    let tray_config_id = tray_config["id"].as_str().unwrap();
-
-    let experiment_payload = serde_json::json!({
-        "name": "Coordinate Mapping Test",
-        "tray_configuration_id": tray_config_id,
-        "is_calibration": false
-    });
-
-    let experiment_response = app
-        .clone()
-        .oneshot(
-            axum::http::Request::builder()
-                .method(axum::http::Method::POST)
-                .uri("/api/experiments")
-                .header("content-type", "application/json")
-                .body(axum::body::Body::from(experiment_payload.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let experiment_body = axum::body::to_bytes(experiment_response.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let experiment: Value = serde_json::from_slice(&experiment_body).unwrap();
-    let experiment_id = experiment["id"].as_str().unwrap();
-
-    let _upload_result = upload_excel_file(&app, experiment_id).await;
-
-    // Fetch results and validate coordinate mappings
-    let results_response = app
-        .clone()
-        .oneshot(
-            axum::http::Request::builder()
-                .method(axum::http::Method::GET)
-                .uri(&format!("/api/experiments/{}", experiment_id))
-                .body(axum::body::Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let results_body = axum::body::to_bytes(results_response.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let experiment_with_results: Value = serde_json::from_slice(&results_body).unwrap();
-    let well_summaries = experiment_with_results["results_summary"]["well_summaries"]
-        .as_array()
-        .expect("Should have well summaries");
-
-    // Validate that we have exactly 192 wells with proper coordinates
-    assert_eq!(well_summaries.len(), 192, "Should have exactly 192 wells");
-
-    let mut p1_wells = 0;
-    let mut p2_wells = 0;
-    let mut coordinate_set = std::collections::HashSet::new();
-
-    for well in well_summaries {
-        let tray_name = well["tray_name"].as_str().unwrap_or("unknown");
-        let coordinate = well["coordinate"].as_str().unwrap_or("unknown");
-
-        match tray_name {
-            "P1" => p1_wells += 1,
-            "P2" => p2_wells += 1,
-            _ => panic!("Unexpected tray name: {}", tray_name),
-        }
-
-        // Validate coordinate format (A1-H12)
-        assert!(
-            coordinate.len() >= 2 && coordinate.len() <= 3,
-            "Coordinate {} should be 2-3 characters",
-            coordinate
-        );
-        assert!(
-            coordinate.chars().next().unwrap().is_ascii_uppercase(),
-            "Coordinate {} should start with A-H",
-            coordinate
-        );
-
-        // Add to set to check for duplicates within tray
-        let full_coord = format!("{}_{}", tray_name, coordinate);
-        assert!(
-            coordinate_set.insert(full_coord.clone()),
-            "Duplicate coordinate found: {}",
-            full_coord
-        );
-    }
-
-    assert_eq!(p1_wells, 96, "Should have 96 P1 wells, got {}", p1_wells);
-    assert_eq!(p2_wells, 96, "Should have 96 P2 wells, got {}", p2_wells);
-
-    println!("   ✅ P1 wells: {} ✓", p1_wells);
-    println!("   ✅ P2 wells: {} ✓", p2_wells);
-    println!("   ✅ Unique coordinates: {} ✓", coordinate_set.len());
-    println!("   🗺️  Well coordinate mapping validated successfully");
-}
